docs_requirements.txt
mkdocs >= 1.3.0
mkdocs-material >= 8.3.7
mkdocs-macros-plugin >= 0.7.0
mkdocs-click >= 0.8.0